Deprecate experimental.aero.rotations
