+++ src/archimedes/experimental/aero/__init__.py
@@ -15,21 +15,10 @@ from .atmosphere import (
     StandardAtmosphere1976Config,
     AtmosphereConfig,
 )
-from .rotations import (
-    dcm_from_euler,
-    x_dcm,
-    y_dcm,
-    z_dcm,
-    dcm_from_quaternion,
-    euler_kinematics,
-    euler_to_quaternion,
-    quaternion_derivative,
-    quaternion_inverse,
-    quaternion_multiply,
-    quaternion_to_euler,
-)
 from .rigid_body import (
     RigidBody,
+    RigidBodyConfig,
+    euler_kinematics,
 )
 from .sensors import (
     Accelerometer,
@@ -43,16 +32,7 @@ from .frames import wind_frame
 
 __all__ = [
     "RigidBody",
-    "quaternion_inverse",
-    "quaternion_multiply",
-    "dcm_from_quaternion",
-    "dcm_from_euler",
-    "x_dcm",
-    "y_dcm",
-    "z_dcm",
-    "quaternion_derivative",
-    "euler_to_quaternion",
-    "quaternion_to_euler",
+    "RigidBodyConfig",
     "wind_frame",
     "euler_kinematics",
     "GravityModel",
